Fix things up

+++ db/migrate/20250301151137_remove_fn_fields_from_custom_functions.rb
@@ -1,5 +1,7 @@
 class RemoveFnFieldsFromCustomFunctions < ActiveRecord::Migration[7.0]
   def change
+    return if Rails.env.production?
+
     remove_column :bootcamp_custom_functions, :name
     rename_column :bootcamp_custom_functions, :fn_name, :name
     rename_column :bootcamp_custom_functions, :fn_arity, :arity
